Change console.log to console.error and renamed tests to utilities
